Move Transport class to com.vaadin.shared.ui (#2716)
