+++ flow-server/src/main/java/com/vaadin/flow/nodefeature/PushConfigurationMap.java
@@ -19,7 +19,7 @@ import java.util.Collection;
 
 import com.vaadin.flow.StateNode;
 import com.vaadin.shared.communication.PushMode;
-import com.vaadin.shared.ui.ui.Transport;
+import com.vaadin.shared.ui.Transport;
 import com.vaadin.ui.PushConfiguration;
 
 /**
